sl-segmentation/requirements.txt
absl-py==1.2.0
pose_format==0.2.2
numpy==1.23.2
pympi-ling==1.70.2
matplotlib==3.7.1
pandas==2.0.1
